Add authentication, login support (via JWT), logging
Users can now create accounts via UsersController and then login via
LoginController. When logging in a JWT is returned. This needs to be
saved by the client and sent as a bearer token for authentication.

Logging was also configured in some places to help debugging.

The security module contains all logic for hashing / checking passwords.