Make send-events-methods as extensions, remove unused lobby, refactor
Previously sending events from the gamehub was created in each method.
This could result in one Event having a different receiving interface
on the frontend. Instead all events are sent by methods defined as
extensions of IClientProxy, which enforces the interface.

Removed unused lobby functionality as that didn't align with the
frontend.

Removed all catches of authentication error as that is resolved by
asp.net functionality under the hood.

Refactored some methods.